implement calculatePredictedMu()

+++ source/extensions/load_balancing_policies/least_response_time/least_response_time_lb.h
@@ -132,7 +132,7 @@ private:
   static double calculateAveRtt(double lambda, double mu, u_int32_t c);
   static u_int64_t calculateCSsthresh(u_int64_t previous_c_ssthresh, double rho, double target_rho);
   static u_int32_t calculatePredictedC(u_int32_t previous_c, double rho, double target_rho, u_int64_t c_ssthresh);
-  static u_int64_t calculatePredictedMu(u_int32_t c, u_int64_t lambda, u_int64_t average_rtt);
+  static u_int64_t calculatePredictedMu(u_int32_t c, double lambda, double average_rtt);
   void updateWeights(const std::vector<HostSharedPtr>& hosts) const;
 
   mutable std::vector<u_int64_t> host_lambdas_;
